requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.7.1
Django==5.0.1
django-allauth==0.60.1
django-countries==7.5.1
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
idna==3.6
oauthlib==3.2.2
phonenumbers==8.13.29
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==69.0.3
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
